Refactored `StartStimulus` - removed boolean argument from StartStimulus method, instead inheriting behaviours that send window markers to new abstract `WindowedControllerBehavior` class - moved window properties and method implementations to `WindowedControllerBehavior` - Wrapped Trial start and end markers in virtual functions, blocking them if relevant through overloads in P300 behavior - removed flag from `BCIController` and instance `StartStimulus` passthroughs - corrected test usage of removed method argument - trimmed and reorganized `BCIControllerBehavior` imports
