+++ Runtime/Scripts/Behaviors/WindowedControllerBehavior.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d5a80028e94cd50488381aab9bd393e3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
